Rename models.py -> model.py; move artifacts inside the expert package
- model.py (was models.py); imports updated in signature/module/dataset + tests.
- compiled artifacts now live under modules/proof_completion/artifacts/ (was
  backend/experts/artifacts/) so each expert is fully self-contained; .gitignore
  updated to backend/experts/modules/*/artifacts/, optimize --out default moved.
- drop unused import in module.py; README layout updated.

46 tests pass.
